+++ qcp/algorithms/__init__.py
