+++ qcp/algorithms/abstract_algorithm.py
